fix: Address all Copilot review comments
Database:
- Fix RoomDatabase.Callback to use SupportSQLiteDatabase signature
- Remove GlobalScope usage, use proper thread-safe initialization
- Make INSTANCE public for synchronous access in callback

Repository:
- Fix integer overflow in nextReviewTime calculation by using Long

UI:
- Apply color parameter to ReviewButton containers (Again/Hard/Good/Easy colors)
- Add delete affordance to DeckListItem (delete button)
- Show 'Deck not found' instead of indefinite spinner in DeckDetailScreen
- Use stringResource() for localization instead of hard-coded strings
- Add 'delete' string resource

Resources:
- Add launcher icon definitions (ic_launcher and ic_launcher_round)
- Fix data_extraction_rules.xml to use proper cloud-backup schema
- Update .gitignore with note about gradle-wrapper.jar auto-download

All review issues resolved and app should now compile successfully.

+++ app/src/main/java/com/catalanflashcard/data/repository/FlashcardRepository.kt
@@ -54,7 +54,7 @@ class FlashcardRepository(
             interval = newInterval,
             easeFactor = newEaseFactor,
             repetitions = newRepetitions,
-            nextReviewTime = System.currentTimeMillis() + (newInterval * 86400000),
+            nextReviewTime = System.currentTimeMillis() + (newInterval.toLong() * 86400000L),
             updatedAt = System.currentTimeMillis()
         )
         cardDao.update(updatedCard)
